% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_regulatory_psites}
\alias{benchmark_regulatory_psites}
\title{benchmark_regulatory_psites}
\usage{
benchmark_regulatory_psites(regulatory_psites)
}
\arguments{
\item{sif}{regulatory_psites, e.i. the first element of the list returned by
regulatory_psites function}
}
\value{
a table comparing the predicted regulatory psites with the regulatory
psites of phosphositeplus
}
\description{
This function check the coherence of regulatory phosphosites predicted by
phonemes with ground truth from phosphositeplus
}
