% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_carnival_options.R
\name{check_carnival_options}
\alias{check_carnival_options}
\title{Check CARNIVAL Options}
\usage{
check_carnival_options(opts)
}
\arguments{
\item{opts}{An object of type \dQuote{\code{list}} defining the run parameters
for CARNIVAL  Use the \code{\link{default_carnival_options}} function to
create a list with default parameter settings. If cplex or cbc are chosen
as the solver, the parameter solverPath needs to be supplied (not
automatically added by \code{default_carnival_options()}).}
}
\description{
Checks the list of input options for CARNIVAL for completeness.
}
\seealso{
\code{\link{default_carnival_options}},
\code{\link[PHONEMeS]{run_phonemes}}
}
