% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_phonemes.R
\name{get_protein_network}
\alias{get_protein_network}
\title{get_protein_network}
\usage{
get_protein_network(phonemes_res)
}
\arguments{
\item{phonemes_res}{phonemes result from the run_phonemes function}
}
\value{
Phonemes network only consisting of protein-protein interactions
}
\description{
This function readd links between phosphosite and their correpsonding proteins
}
