\name{PHONEMeS-package}
\alias{PHONEMeS-package}
\alias{PHONEMeS}
\docType{package}
\title{
MS phosphoproteomics-based models of signalling
}
\description{
This package implements the method in Terfve et al. (submitted) to build logic-based models of signalling networks based on mass spectrometry phosphoproteomics data and kinase/phosphatase-substrate relationships.
}
\details{
\tabular{ll}{
Package: \tab PHONEMeS\cr
Type: \tab Package\cr
Version: \tab 0.2.6\cr
Date: \tab 2015-02-14\cr
License: \tab GPL-2\cr
}
}
\author{
C Terfve
Maintainer: C Terfve <terfve@ebi.ac.uk>
}