% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{phonemesKSN}
\alias{phonemesKSN}
\title{The KSN used as prior knowledge network when running vanilla Phonemes.
The KSN contains kinase-substrate interactions via a phosphorylation site}
\format{
The full PKN contains 37,631 kinase-substrate interactions via a phosphorylation site (1,710 kinases; 15,413 phosphosites)
\describe{
\item{source}{upstream kinase/phosphorylation site}
\item{interaction}{connection (1)}
\item{target}{downstream phosphorylation site/kinase}
}
}
\source{
\url{https://github.com/saezlab/PHONEMeS/blob/master/data-raw/phonemesKSN.R}
}
\description{
The KSN used as prior knowledge network when running vanilla Phonemes.
The KSN contains kinase-substrate interactions via a phosphorylation site
}
\examples{
data("phonemesKSN")
}
\keyword{KSN}
