% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{process_phosphositeplus_regpsites}
\alias{process_phosphositeplus_regpsites}
\title{process_phosphositeplus_regpsites}
\usage{
process_phosphositeplus_regpsites(ppsp_regpsites_file_path)
}
\arguments{
\item{regpsite_file}{path to the tab file of regulatory psites of phosphositeplus}
}
\value{
a formatted table of regulatory psites of phosphositeplus
with their known mode of regulation
}
\description{
This function process the tab file of regulatory psites of phosphositeplus,
downloaded from https://www.phosphosite.org/staticDownloads
}
