% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_phonemes.R
\name{reattach_psites}
\alias{reattach_psites}
\title{Reattach_psites}
\usage{
reattach_psites(phonemes_res)
}
\arguments{
\item{phonemes_res}{phonemes result from the run_phonemes function}
}
\value{
List of PHONEMES results and final inputObj, measObj, netObj used, with psites attached
}
\description{
This function readd links between phosphosite and their correpsonding proteins
}
