% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{regulatory_psites}
\alias{regulatory_psites}
\title{regulatory_psites}
\usage{
regulatory_psites(phonemes_res, th_act = 0)
}
\arguments{
\item{phonemes_res}{carnival result from the run_carnival function}

\item{th_act}{threshold for node activity to include in the benchmark}
}
\value{
list with two elements: the regulatory psites with predicted mode of
regulations and the kinases that catalyse their phosphorilation
}
\description{
This function  extract regulatory phosphites from the phonemes network, e.i.
psites that are differentially regulated on proteins that are found in
the phonemes network
}
