% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_phonemes.R
\name{run_phonemes}
\alias{run_phonemes}
\title{Run PHONEMeS}
\usage{
run_phonemes(
  inputObj,
  measObj,
  netObj = phonemesPKN,
  rmNodes = NULL,
  pruning = TRUE,
  n_steps_pruning = 50,
  solverPath,
  solver = "cplex",
  timelimit = 7200,
  mipGAP = 0.05,
  poolrelGAP = 0
)
}
\arguments{
\item{inputObj}{named vector of perturbation targets. Either 1 (up regulated) or -1 (down regulated)}

\item{measObj}{named vector of the measurements}

\item{netObj}{data frame of the prior knowledge network}

\item{rmNodes}{character vector of nodes to remove from prior knowledge network}

\item{pruning}{logic, set to TRUE if network should be pruned (recommended)}

\item{n_steps_pruning}{integer giving the order of the neighborhood}

\item{solverPath}{path to the solver}

\item{solver}{one of the solvers available from getSupportedSolvers()}

\item{timelimit}{solver time limit in seconds}

\item{mipGAP}{CPLEX parameter: absolute tolerance on the gap}

\item{poolrelGAP}{CPLEX/Cbc parameter: Allowed relative gap of accepted}
}
\value{
List of CARNIVAL results and final inputObj, measObj, netObj used
}
\description{
This function runs CARNIVAL with the input of phosphoproteomic data (phosphosites and kinases).
The prior knowledge network used is the combination of protein-protein and protein-phosphosite
interactions from omnipath. Before running CARNIVAL the network is pruned by removing nodes n_steps
upstream and downstream of measurements and inputs, respectively.
}
