% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanilla_phonemes.R
\name{run_vanilla_phonemes}
\alias{run_vanilla_phonemes}
\title{Run vanilla PHONEMeS (PHONEMeS v1.0.0)}
\usage{
run_vanilla_phonemes(
  inputObj,
  measObj,
  netObj = phonemesKSN,
  rmNodes = NULL,
  pruning = FALSE,
  n_steps_pruning = 50,
  solverPath,
  solver = "cplex",
  timelimit = 7200,
  mipGAP = 0.05,
  poolrelGAP = 0
)
}
\arguments{
\item{inputObj}{named vector of perturbation targets}

\item{measObj}{named vector of the measurements}

\item{netObj}{data frame of the prior knowledge network}

\item{rmNodes}{character vector of nodes to remove from prior knowledge network}

\item{pruning}{logic, set to TRUE if network should be pruned (default = FALSE)}

\item{n_steps_pruning}{integer giving the order of the neighborhood}

\item{solverPath}{path to the solver}

\item{solver}{one of the solvers available from getSupportedSolvers()}

\item{timelimit}{solver time limit in seconds}

\item{mipGAP}{CPLEX parameter: absolute tolerance on the gap}

\item{poolrelGAP}{CPLEX/Cbc parameter: Allowed relative gap of accepted}
}
\value{
List of CARNIVAL results and final inputObj, measObj, netObj used
}
\description{
The code for this function is comparable to the orginal PHONEMeS package. An unsigned
kinase-substrate network is used as prior knowledge connecting perturbed kinases with phosphosites.
}
